Feature/board

+++ src/main/java/Team/project/itda/Service/BoardServiceImpl.java
@@ -4,13 +4,19 @@ import Team.project.itda.DTO.BoardDTO;
 import Team.project.itda.DTO.BoardPageRequestDTO;
 import Team.project.itda.DTO.BoardPageResultDTO;
 import Team.project.itda.Entity.Board;
+import Team.project.itda.Entity.QBoard;
+import Team.project.itda.Entity.UserEntity;
 import Team.project.itda.Repository.BoardRepository;
+import com.querydsl.core.BooleanBuilder;
+import com.querydsl.core.types.dsl.BooleanExpression;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.security.access.AccessDeniedException;
+
 
 import java.util.Optional;
 import java.util.function.Function;
@@ -23,12 +29,12 @@ public class BoardServiceImpl implements BoardService{
     private final BoardRepository repository;
 
     @Override
-    public Long register(BoardDTO dto) {
+    public Long register(BoardDTO dto, UserEntity user) {
 
         log.info("DTO-----------");
         log.info(dto);
 
-        Board entity = dtoToEntity(dto);
+        Board entity = dtoToEntity(dto, user);
 
         log.info(entity);
 
@@ -41,7 +47,9 @@ public class BoardServiceImpl implements BoardService{
     public BoardPageResultDTO<BoardDTO, Board> getList(BoardPageRequestDTO requestDTO) {
         Pageable pageable = requestDTO.getPageable(Sort.by("bno").descending());
 
-        Page<Board> result = repository.findAll(pageable);
+        BooleanBuilder booleanBuilder = getSearch(requestDTO);
+
+        Page<Board> result = repository.findAll(booleanBuilder, pageable);
 
         Function<Board, BoardDTO> fn = (entity ->
                 entityToDto(entity));
@@ -56,21 +64,70 @@ public class BoardServiceImpl implements BoardService{
     }
 
     @Override
-    public void remove(Long bno) {
-        repository.deleteById(bno);
-    }
+    public void remove(Long bno, UserEntity user) {
+        Optional<Board> result = repository.findById(bno);
 
-    @Override
-    public void modify(BoardDTO dto) {
-        Optional<Board> result = repository.findById(dto.getBno());
         if (result.isPresent()) {
             Board entity = result.get();
 
+            if (!entity.getWriter().equals(user)) {
+                throw new AccessDeniedException("dd");
+            }
+            repository.deleteById(bno);
+        }
+    }
+
+    @Override
+    public void modify(BoardDTO dto, UserEntity user) {
+        Optional<Board> result = repository.findById(dto.getBno());
+
+        if (result.isPresent()) {
+            Board entity = result.get();
+
+            if (!entity.getWriter().equals(user)) {
+                throw new AccessDeniedException("dd");
+            }
             entity.changeTitle(dto.getTitle());
             entity.changeCategory(dto.getCategory());
             entity.changeContent(dto.getContent());
 
             repository.save(entity);
         }
+    }
+
+    private BooleanBuilder getSearch(BoardPageRequestDTO requestDTO) {
+        String type = requestDTO.getType();
+
+        BooleanBuilder booleanBuilder = new BooleanBuilder();
+
+        QBoard qBoard = QBoard.board;
+
+        String keyword = requestDTO.getKeyword();
+
+        BooleanExpression expression = qBoard.bno.gt(0L); // bno > 0
+
+        booleanBuilder.and(expression);
+
+        if (type == null || type.trim().length() == 0) {
+            return booleanBuilder;
+        }
+
+        BooleanBuilder conditionBuilder = new BooleanBuilder();
+
+        if (type.contains("t")) {
+            conditionBuilder.or(qBoard.title.contains(keyword));
+        }
+        if (type.contains("c")) {
+            conditionBuilder.or(qBoard.content.contains(keyword));
+        }
+        if (type.contains("w")) {
+            conditionBuilder.or(qBoard.writer.username.contains(keyword));
+        }
+
+
+        booleanBuilder.and(conditionBuilder);
+
+        return booleanBuilder;
+
     }
 }
